pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
Flask = "^2.2.2"
Flask-Cors = "^3.0.10"
python-dotenv = "^0.21.0"
pytest = "^7.2.1"
requests = "^2.28.2"
flask-mysql = "^1.5.2"
pytest-cov = "^4.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
flake8-annotations = "^3.0.0"
flake8-docstrings = "^1.7.0"
pre-commit = "^3.0.2"
cryptography = "^39.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
